misc/HtmlTestSuite/results/test-0272.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Multi-column Layout: parsing column-rule-color with invalid values</title>
<link href="https://drafts.csswg.org/css-multicol/#propdef-column-rule-color" rel="help"/>
<meta content="column-rule-color supports only the grammar '&lt;color&gt;'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_invalid_value("column-rule-color", "auto");

test_invalid_value("column-rule-color", "green blue");
test_invalid_value("column-rule-color", "green, blue");
</script>
</body>
</html>
